auth-gateway/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
azure-functions==1.23.0
bcrypt==4.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
idna==3.10
MarkupSafe==3.0.2
motor==3.7.0
passlib==1.7.4
pyasn1==0.4.8
pydantic==2.11.4
pydantic_core==2.33.2
pymongo==4.12.1
python-jose==3.4.0
python-multipart==0.0.20
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
